KeyBindingMixin: implement "tap to toggle"
We want to only toggle to on if the player briefly taps the key.  That
way if the player holds down "W" to walk around for a few second it
won't count as a tap, WON'T toggle it on, and when the player releases
the key the PC will stop walking.